feat: scaffold oms-validator service and refactor to BigDecimal
- Add oms-validator module to parent POM
- Scaffold complete oms-validator service:
  - Kafka consumer for order.ingest topic
  - Order validation (params, symbol, risk checks)
  - Risk checks (buying power, order value, position limits)
  - Status updates (VALIDATED/REJECTED)
  - Publisher to order.validated/order.rejected topics
  - Comprehensive README with architecture docs
- Refactor numeric fields to BigDecimal:
  - OrderDTO: quantity, filledQuantity, limitPrice, stopPrice
  - Order entities in both oms-ingest and oms-validator
  - OrderIngestionService conversions
  - Validator risk check calculations
- Fix dependency issues:
  - Exclude kafka-server from spring-kafka-test (doesn't exist in Maven Central)
  - Use tools.jackson.* (Jackson 3) consistently
- Database schema already uses DECIMAL(18,6) - no migration needed
- Both services compile and package successfully

+++ shared/common-models/src/main/java/com/oms/common/model/OrderDTO.java
@@ -1,5 +1,6 @@
 package com.oms.common.model;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -44,13 +45,13 @@ public class OrderDTO {
     private OrderType orderType;
 
     @Positive(message = "Quantity must be positive")
-    private Double quantity;
+    private BigDecimal quantity;
 
-    private Double filledQuantity;
+    private BigDecimal filledQuantity;
 
-    private Double limitPrice;
+    private BigDecimal limitPrice;
 
-    private Double stopPrice;
+    private BigDecimal stopPrice;
 
     @NotNull(message = "Time in force is required")
     private TimeInForce timeInForce;
